refactor: target only v2.1+ releases of laminas-httphandlerrunner
Instead of doing a new minor and targeting both the v1 and v2 series of laminas-httphandlerrunner, we'll instead target just v2, and issue a new major of this library.
