+++ psalm.xml.dist
@@ -15,24 +15,6 @@
     </projectFiles>
 
     <issueHandlers>
-        <ConstructorSignatureMismatch>
-            <errorLevel type="suppress">
-                <file name="src/SwooleRequestHandlerRunner.php"/>
-            </errorLevel>
-        </ConstructorSignatureMismatch>
-
-        <DeprecatedClass>
-            <errorLevel type="suppress">
-                <referencedClass name="Mezzio\Swoole\RequestHandlerRunner\V2RequestHandlerRunner"/>
-            </errorLevel>
-        </DeprecatedClass>
-
-        <ImplementedParamTypeMismatch>
-            <errorLevel type="suppress">
-                <file name="src/SwooleRequestHandlerRunner.php"/>
-            </errorLevel>
-        </ImplementedParamTypeMismatch>
-
         <InternalClass>
             <errorLevel type="suppress">
                 <referencedClass name="Mezzio\Swoole\Log\StrftimeToICUFormatMap"/>
@@ -56,18 +38,6 @@
                 <referencedMethod name="PHPUnit\Framework\MockObject\Builder\InvocationMocker::with"/>
             </errorLevel>
         </InternalMethod>
-
-        <InvalidExtendClass>
-            <errorLevel type="suppress">
-                <referencedClass name="Mezzio\Swoole\SwooleRequestHandlerRunner"/>
-            </errorLevel>
-        </InvalidExtendClass>
-
-        <MethodSignatureMismatch>
-            <errorLevel type="suppress">
-                <file name="src/RequestHandlerRunner/RequestHandlerRunnerTrait.php"/>
-            </errorLevel>
-        </MethodSignatureMismatch>
     </issueHandlers>
 
     <plugins>
