+++ src/Event/ServerStartListenerFactory.php
@@ -10,7 +10,7 @@ namespace Mezzio\Swoole\Event;
 
 use Mezzio\Swoole\Log\AccessLogInterface;
 use Mezzio\Swoole\PidManager;
-use Mezzio\Swoole\RequestHandlerRunner\RequestHandlerConstantsInterface;
+use Mezzio\Swoole\SwooleRequestHandlerRunner;
 use Psr\Container\ContainerInterface;
 use Psr\Log\LoggerInterface;
 use Webmozart\Assert\Assert;
@@ -37,7 +37,7 @@ final class ServerStartListenerFactory
         Assert::string($appRoot);
 
         $processName = $config['swoole-http-server']['process-name']
-            ?? RequestHandlerConstantsInterface::DEFAULT_PROCESS_NAME;
+            ?? SwooleRequestHandlerRunner::DEFAULT_PROCESS_NAME;
         Assert::stringNotEmpty($processName);
 
         return new ServerStartListener($pidManager, $logger, $appRoot, $processName);
