+++ src/Event/WorkerStartListener.php
@@ -8,7 +8,7 @@ declare(strict_types=1);
 
 namespace Mezzio\Swoole\Event;
 
-use Mezzio\Swoole\RequestHandlerRunner\RequestHandlerConstantsInterface;
+use Mezzio\Swoole\SwooleRequestHandlerRunner;
 use Psr\Log\LoggerInterface;
 
 use function chdir;
@@ -32,7 +32,7 @@ class WorkerStartListener
     public function __construct(
         LoggerInterface $logger,
         string $cwd,
-        string $processName = RequestHandlerConstantsInterface::DEFAULT_PROCESS_NAME
+        string $processName = SwooleRequestHandlerRunner::DEFAULT_PROCESS_NAME
     ) {
         $this->logger      = $logger;
         $this->cwd         = $cwd;
