--- src/RequestHandlerRunner/RequestHandlerConstantsInterface.php
@@ -1,13 +1,0 @@
-<?php
-
-declare(strict_types=1);
-
-namespace Mezzio\Swoole\RequestHandlerRunner;
-
-interface RequestHandlerConstantsInterface
-{
-    /**
-     * Default Process Name
-     */
-    public const DEFAULT_PROCESS_NAME = 'mezzio';
-}
